refactor(tauri): finish git command boundary split
