+++ apps/desktop/src-tauri/src/commands/git/authorization/mod.rs
@@ -1,0 +1,11 @@
+mod cache;
+mod listing;
+mod metadata;
+mod resolution;
+
+pub(crate) use cache::{
+    authorized_worktree_cache, cache_key, invalidate_worktree_resolution_cache_for_repo,
+    AuthorizedWorktreeCacheEntry,
+};
+pub(crate) use metadata::{read_git_common_dir, read_worktree_state_token};
+pub(crate) use resolution::resolve_working_dir;
