+++ apps/desktop/src-tauri/src/commands/git/authorization/resolution.rs
@@ -1,0 +1,37 @@
+use std::{fs, path::PathBuf};
+
+use super::cache::is_authorized_worktree;
+
+pub(super) fn canonicalize_for_validation(path: &str, field: &str) -> Result<PathBuf, String> {
+    fs::canonicalize(path)
+        .map_err(|_| format!("{field} does not exist or is not accessible: {path}"))
+}
+
+/// Resolve the effective path for a git operation. If `working_dir` is
+/// provided, it is validated as a git worktree/repo and used instead of
+/// `repo_path`. The caller must have already authorized `repo_path`.
+pub(crate) fn resolve_working_dir(
+    repo_path: &str,
+    working_dir: Option<&str>,
+) -> Result<String, String> {
+    let canonical_repo = canonicalize_for_validation(repo_path, "repo_path")?;
+
+    match working_dir {
+        Some(wd) if !wd.is_empty() && wd != repo_path => {
+            let canonical_working_dir = canonicalize_for_validation(wd, "working_dir")?;
+
+            if canonical_working_dir == canonical_repo {
+                return Ok(canonical_working_dir.to_string_lossy().to_string());
+            }
+
+            if is_authorized_worktree(canonical_repo.as_path(), canonical_working_dir.as_path())? {
+                return Ok(canonical_working_dir.to_string_lossy().to_string());
+            }
+
+            Err(format!(
+                "working_dir is not within authorized repository or linked worktrees: {wd}"
+            ))
+        }
+        _ => Ok(canonical_repo.to_string_lossy().to_string()),
+    }
+}
