+++ apps/desktop/src-tauri/src/commands/git/tests/authorization/mod.rs
@@ -1,0 +1,3 @@
+mod cache_invalidation;
+mod resolution;
+mod state_token;
