+++ apps/desktop/src-tauri/src/commands/git/tests/fixtures/mod.rs
@@ -1,0 +1,18 @@
+mod app;
+mod git_port;
+mod repo;
+mod task_store;
+
+pub(super) use app::{
+    invoke_json, setup_command_git_fixture, setup_command_git_fixture_with_mutations,
+    setup_command_git_fixture_with_summary,
+};
+pub(super) use git_port::{
+    WorktreeStatusCall, WorktreeStatusResult, WorktreeStatusSummaryCall,
+    WorktreeStatusSummaryResult,
+};
+pub(super) use repo::{
+    clear_authorized_worktree_cache_for_repo, init_repo, run_git, sample_worktree_status_data,
+    sample_worktree_status_summary_data, seed_authorized_worktree_cache_with_subset,
+    unique_test_dir,
+};
